Improve javadocs across Lorenz
This commit introduces a lot of author and since tags. I will obliterate since tags on members when I release 1.0.0. I want 0.3.0 to be a huge improvement for javadocs, where Lorenz has not lacked, but never really been fleshed out. Introducing Enigma+JAM support, though the base has begun, will be for the 0.4.0 release.

I will be releasing 0.3.0 sometime in the next week, it will be a largely cosmetic update with little functional changes.