feat: Initial work on Cadix
Cadix is (maybe) going to be the future of IO in Lorenz, and will ultimately be made separate from Lorenz. It makes
use of the visitor patterm, which is rather nice for this sort of work.

Though this is not a set-in-stone change that I am going to make, I may well end up dropping this - or pushing it beyond
0.4.0.

Should this be completed, Cadix will be modularised in its own repository:

- cadix-core
- cadix-srg
- cadix-csrg
- cadix-tsrg

Cadix may also help with supporting formats such as Enigma and JAM.