Add new improved and extensible merge API
This commit maintains backwards compatibility with the previous merge
methods, but implements the merge functionality separately from the
mappings. The old merge methods now call out to the new API with the
default implementation instead.

The goal of this commit was two-fold:

 1. Implement a more robust mapping merge system which can handle more
    unique situations correctly, including (and especially) cases where
    incomplete mapping sets are merged.
 2. Provide an extensible API that callers can use to customize the
    behavior of the merge process. Provide multiple layers of
    extensibility for simpler and more complex use-cases both supported.

This is accomplished with 2 new interfaces:

 1. MappingSetMerger
 2. MappingSetMergerHandler

The first class, MappingSetMerger, is the API users will interact with.
It provides the user-facing methods for merging MappingSets and
individual mappings, but it also serves as the controller for the second
class.

MappingSetMergeHandler is a fully-default interface which is used by the
default implementation of MappingSetMerger. It has methods for each
individual mapping case which can arrise, allowing  implementers to
override specifically the cases they wish to modify the behavior of. The
default implementation is suitable for a typical merge, but this API
allows the merge API to be used for many more operations.

The code in  MappingSetMerger is considerably more complex than
MappingSetMergerHandler because of this. MappingSetMerger spots the
different possible mapping cases and calls out to the relevant methods,
allowing the implementer of MappingSetMergerHandler to have a much
simpler view of the mapping proces, and hopefully simplify code.

All of this and more is also described in extensive detail in Javadocs
of the relevant classes.

This commit also fixes a couple Javadoc warnings and sets the javadoc
build up to support the now standard @apiNote, @implNote, and  @implSpec
Javadoc tags.