Allow any method parameter index
The previous implementation required all method parameter indexes be
within the range for the parameters of the method mapping based on the
method signature. This commit changes the behavior to instead allow any
parameter index provided.

There are two arguments for allowing this behavior:

 1. Some mapping formats have different meanings for the index
    parameters, and Lorenz has no way of handling them other than taking
    the parameter index as-is. The most common example is static methods
    starting to count method parameters at 0, while instance methods
    start at 1 (as `this` is index 0). Lorenz has no way of knowing a
    method is static or not, and even if it did it can't be guaranteed
    that every mapping format behaves the same way.

 2. All other mappings (top-level class, field, method, and inner class
    mappings) have no way to verify, and don't verify, if the mapping is
    valid. A method parameter index outside of the parameter index range
    is very similar to a field mapping in a class which doesn't contain
    a field of that name. It's Lorenz's job to verify mappings are
    valid, only to store them as given and reproduce them later.

There's definitely a risk here that 1 mapping format may want 0-indexed
parameter mappings for all methods, whereas another mapping format wants
the above strategy of using 1-indexed mappings for instance methods, but
I don't see a viable solution for handling cases like these.

In situations where parameter mappings are needed it's likely that the
mapping format will not change, or if it does then it is up to the user
to handle it one way or the other. Lorenz is just a mapping container,
it can't perfectly handle all cases. But as long as it provides a
flexible model and allows the user to insert the mappings it needs,
then the user can write whatever code they need to to handle whichever
situation they run into.