Declutter changelogs
Though we may decide to change this somewhat in the future, this is a
good step to clean up the ever-growing directory listing.

With the commit, changelogs for published versions reside in
'changelogs/x.y/x.y.z.md'. Currently work-in-progress versions still
sit within `changelogs` - this is just to promote the approriate and
current changelog.